refactor(react): extract individual react elements to files
